Update SimpleLinearRegression.java
Refactoring constructor to take in a CSVParser object rather than instantiating in the SimpleLinearRegression class. Since I was trying to mock 'CSVParse' in my test class and the object itself is being instantiated in the class that I was trying to test, I inevitably ran into some issues. The primary problem is that the object cannot be mocked because it is being recreated in during the test.